Optimization.
Added a temporary variable for -D_i D_j alpha + alpha R_ij and it's trace
in order to avoid recomputing this when making it trace-free.

Also shortened the calculation of the Christoffel symbols from the
conformal christoffel symbols. Now they don't involve the derivatives
of the determinant of the physical metric, but only derivatives of phi.

The number of lines in the RHS calculation went down from 1240 to 1203.


Signed-off-by: Peter Diener <diener@10-4-34-85.lsu.edu>