Remove superfluous synchronisation of RHS variables.
Calculate ADM variables in two parts, so that only the time derivatives of lapse and shift are synchronised.

Introduce a new parameter "dt_lapse_shift_method" that can be used to omit out the advection terms there, so that the time derivatives of lapse and shift do not need to be synchronised either.  (This may be physically wrong.)