Move RHS evaluations around to optimize for 32Kb instruction cache machines.
Move RHS evaluations for trK and the lapse to RHS1 to better even out the
instruction count between RHS1 and RHS2 leading to better performance on
32Kb instruction cache machines. Also replace covariant derivatives with
respect to the physical metric with covariant derivatives with respect to
the conformal metric. This avoids having to compute the physical Christoffel
symbols. The performance on 64Kb instruction cache machines is unchanged.
Also get rid of some unused shorthands and generate code for 3 timelevels.

Signed-off-by: Peter Diener <diener@cct.lsu.edu>