ML_BSSN_Test: Remove testsuite output for constraints
The current Cactus testsuite mechanism does not allow you to set per-variable absolute and relative tolerances.  The constraints are typically very small in comparison to the evolved variables used to compute them.  This means that roundoff errors in the evolved variables can lead to large relative errors in the constraints.  The correct solution would be to specify only an absolute tolerance for the constraints, but this is not possible at the moment.  In future, a separate test could be written to test that the constraints are being correctly computed on constraint-violating initial data.  For now, we omit the tests of the constraints from the testsuites.