ML_BSSN_Test: Update testsuites to use 4th order differencing in Exact
The Exact thorn uses finite differencing to compute the extrinsic
curvature rather than using the analytic result.  The step size used
is independent of the Cactus grid.  Using a very small step size for
accuracy can lead to the subtraction of floating point numbers which
are very close to each other, leading to a large loss of precision.
This loss of precision can amplify differences between compilers and
optimisation settings.

A recent Exact commit implements 4th order differencing and a
configurable step size.  Hence for a given finite differencing error,
a larger step size can be used, reducing the loss of precision.

This commit to the McLachlan testsuites (parameter files and data)
implements this change by enabling 4th order differencing and a step
size of 2e-4.  A convergence test in the step size has shown that this
leads to a finite differencing error of ~1e-14.  It also leads to more
consistent results across compilers and optimisation settings.  The
data was generated with gcc using -O2 and the testsuites pass with
the Intel compiler also using -O2.

This testsuite data also takes into account the modification to the
constraint boundary condition in the previous commit.