Set RHS variables to zero in ANALYSIS before computing them
The RHS variables can only be computed on the interior in ANALYSIS, so initialise the entire grid to zero before computing the variable.  An alternative would be to create a new looping method "Exterior" which looped over all the non-interior points.  That would avoid setting the points twice, though setting to zero is inexpensive so this is not really necessary.