Improve finite differencing in tests
For those tests which use Exact, the extrinsic curvature is computed using finite differencing.  With a small timestep, this is very sensitive to roundoff.  The timestep can be increased by using fourth order finite differencing, which we do here.  In order to form robust regression tests which give consistent results on different machines, we also run with an unphysically large timestep in Exact.  This is OK for a regression test, but note that the constraint violations will not converge away with the evolution grid spacing.