Correct definition of dot[A] advection terms
There terms were implemented wrong, probably when the advection terms were split out of the main RHS routine. Without this correction, Kerr-Schild is not manifestly stationary any more.

Some of these terms are in code paths that are (to my knowledge) unused, and so it is not clear what to do. I've tried to make these code paths as similar to the others as possible.